feat: Version 1 ready which selects and deletes

+++ java/label-tool/src/io/github/shahrukhqasim/ties/label/Selectable.java
@@ -8,9 +8,8 @@ import javafx.geometry.Rectangle2D;
 public interface Selectable extends Drawable {
     /**
      * If the user tries to select it, the selection area is passed to this function
-     *
-     * @param selectionArea The area user selects
+     *  @param selectionArea The area user selects
      * @param scale Scale (zoom level) selected by the user
      */
-    void select(Rectangle2D selectionArea, float scale);
+    void select(Rectangle2D selectionArea, double scale);
 }
